Test multiple file writes from logger
